move to src/ffi.rs
